requirements.txt
# requirements.txt
# Lista de dependências do projeto

pandas==1.3.3
python-dotenv==0.19.2
chardet==5.1.0
PyPDF2==3.0.0